requirements.txt
dhooks
psutils
python-dotenv
pytz